exports, but it's not a real excel file. It's html There are other options out there, like https://github.com/arydjmal/to_xls/issues/2 but they seem to suffer from the same symptoms.
Option 1: Export as tab-delimited text, and
tell your users to open it with excel (no warnings)

Option 2: Export as whatever you want, but
use an .xls file extension. excel 2007 and
later will warn you that the file is of a
different mime type and may be corrupt.

Which is better?